add comment-details

+++ backend/src/routes/private/food-details.js
@@ -7,6 +7,7 @@ const { handleAPIResponse } = require('../../common/handleAPIResponse');
 router.route('/api/restaurants/:restaurant_id/foods/:food_id')
   .get(validateCustomer, async (req, res, next) => {
     const { restaurant_id, food_id } = req.params;
+    if (!restaurant_id || !food_id) return handleAPIResponse(res, 400, 'restaurant_id && food_id required');
     try {
       const food = await knex('foods').first().where({ restaurant_id, id: food_id });
       return handleAPIResponse(res, 200, food);
@@ -15,7 +16,10 @@ router.route('/api/restaurants/:restaurant_id/foods/:food_id')
     }
   }).put(validateCustomer, async (req, res, next) => {
     const { restaurant_id, food_id } = req.params;
+    if (!restaurant_id || !food_id) return handleAPIResponse(res, 400, 'restaurant_id && food_id required');
     try {
+      let food = await knex('foods').first().where({ restaurant_id, id: food_id });
+      if (!food) return handleAPIResponse(res, 404, 'food_id not exist');
       //
       const update_data = {};
       ['name', 'price', 'status', 'img_url', 'description'].forEach(param => {
@@ -24,7 +28,7 @@ router.route('/api/restaurants/:restaurant_id/foods/:food_id')
       await knex('foods')
         .update(update_data)
         .where({ restaurant_id, id: food_id });
-      const food = await knex('foods').first().where({ restaurant_id, id: food_id });
+      food = await knex('foods').first().where({ restaurant_id, id: food_id });
       return handleAPIResponse(res, 200, food);
     } catch (e) {
       next(e);
@@ -32,7 +36,12 @@ router.route('/api/restaurants/:restaurant_id/foods/:food_id')
   })
   .delete(validateCustomer, async (req, res, next) => {
     const { restaurant_id, food_id } = req.params;
+    if (!restaurant_id || !food_id) return handleAPIResponse(res, 400, 'restaurant_id && food_id required');
+
     try {
+      const food = await knex('foods').first().where({ restaurant_id, id: food_id });
+      if (!food) return handleAPIResponse(res, 404, 'food_id not exist');
+
       //
       await knex('foods').delete().where({ restaurant_id, id: food_id });
       return handleAPIResponse(res, 200);
